fix: followup questions
